extract hmr implementation
